requirements.txt
#general modules
Django==4.2.5
psycopg2-binary==2.9.6
python-decouple==3.8
djangorestframework==3.14.0

#third party modules
django-filter==23.2
Markdown==3.4.4
drf-spectacular
drf-spectacular[sidecar]
djangorestframework-simplejwt==5.3.0

# Testing
pytest
pytest-django

#Refactoring & Test
black
flake8